# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  constant.ProviderUser:
    enum:
    - internal
    - google
    - facebook
    type: string
    x-enum-varnames:
    - PROVIDER_USER_INTERNAL
    - PROVIDER_USER_GOOGLE
    - PROVIDER_USER_FACEBOOK
...
  user_do.CreateUserReq:
    properties:
      email:
        example: leannon.herminia@bxt.com
        maxLength: 255
        minLength: 1
        type: string
      full_name:
        example: Mr. Cortez Nicolas II
        maxLength: 255
        minLength: 1
        type: string
      password:
        example: 9HZdzEr.uWct
        maxLength: 255
        minLength: 1
        type: string
      username:
        example: carroll.myrtle
        maxLength: 255
        minLength: 1
        type: string
    type: object
  user_do.Token:
    properties:
      accessToken:
        example: AccessToken
        type: string
      refreshToken:
        example: RefreshToken
        type: string
    type: object
  user_do.UpdateUserReq:
    properties:
      email:
        example: leannon.herminia@bxt.com
        maxLength: 255
        minLength: 1
        type: string
      full_name:
        example: Mr. Cortez Nicolas II
        maxLength: 255
        minLength: 1
        type: string
      id:
        type: string
      password_hash:
        example: 9HZdzEr.uWct
        maxLength: 255
        minLength: 1
        type: string
      username:
        example: carroll.myrtle
        maxLength: 255
        minLength: 1
        type: string
    type: object
  user_do.UserResponse:
    properties:
      created_time:
        example: 1005708562
        type: integer
      email:
        example: leannon.herminia@bxt.com
        type: string
      full_name:
        example: Mr. Cortez Nicolas II
        type: string
      id:
        example: 01HFGRZQHPY8JHVRZ7NSRDGVB5
        type: string
      provider:
        allOf:
        - $ref: '#/definitions/constant.ProviderUser'
        example: internal
      role:
        example: pjOh1Jty
        type: string
      token:
        $ref: '#/definitions/user_do.Token'
      updated_time:
        example: 1005708562
        type: integer
      username:
        example: carroll.myrtle
        type: string
    type: object
  v_api.Paging:
    properties:
      current_offset:
        description: for load more
        type: integer
      current_page:
        description: for page
        type: integer
      limit:
        description: for load more
        type: integer
      next_offset:
        description: for load more
        type: integer
      next_page:
        description: for page
        type: integer
      page_size:
        description: for page
        type: integer
      prev_page:
        description: for page
        type: integer
      total:
        description: total items
        type: integer
      total_pages:
        description: for page
        type: integer
    type: object
...
  v_api.Response-array_user_do_UserResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/user_do.UserResponse'
        type: array
      data_extend: {}
      description:
        type: string
      encryption:
        type: string
      message:
        type: string
      paging:
        $ref: '#/definitions/v_api.Paging'
      status:
        type: integer
    type: object
...
  v_api.Response-string:
    properties:
      data:
        type: string
      data_extend: {}
      description:
        type: string
      encryption:
        type: string
      message:
        type: string
      paging:
        $ref: '#/definitions/v_api.Paging'
      status:
        type: integer
    type: object
  v_api.Response-user_do_UserResponse:
    properties:
      data:
        $ref: '#/definitions/user_do.UserResponse'
      data_extend: {}
      description:
        type: string
      encryption:
        type: string
      message:
        type: string
      paging:
        $ref: '#/definitions/v_api.Paging'
      status:
        type: integer
    type: object
  v_proto.VolioRpcError:
    properties:
      code:
        type: integer
      details:
        items:
          $ref: '#/definitions/v_proto.VolioRpcErrorDetail'
        type: array
      error_code:
        type: integer
      message:
        type: string
      wait_for:
        type: integer
    type: object
  v_proto.VolioRpcErrorDetail:
    properties:
      field:
        type: string
      msg:
        type: string
      type:
        type: string
    type: object
info:
  contact:
    email: admin@volio.vn
    name: admin
  description: Tài liệu kết nối Test API
  license:
    name: GNU 3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  termsOfService: empty
  title: Test API
  version: "3.0"
paths:
...
  /admin/users:
    delete:
      consumes:
      - application/json
      description: Delete an existing User
      parameters:
      - default: access_token
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v_api.Response-string'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v_proto.VolioRpcError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v_proto.VolioRpcError'
      summary: Delete User
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update an existing User
      parameters:
      - default: access_token
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Update User request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user_do.UpdateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v_api.Response-user_do_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v_proto.VolioRpcError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v_proto.VolioRpcError'
      summary: Update User
      tags:
      - User
  /admin/users/detail:
    get:
      consumes:
      - application/json
      description: Get list of User with optional filters
      parameters:
      - default: access_token
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of the User
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v_api.Response-user_do_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v_proto.VolioRpcError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v_proto.VolioRpcError'
      summary: Get User
      tags:
      - User
  /admin/users/pagination:
    get:
      consumes:
      - application/json
      description: Get pagination of User
      parameters:
      - default: access_token
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: page_size
        type: integer
      - in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v_api.Response-array_user_do_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v_proto.VolioRpcError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v_proto.VolioRpcError'
      summary: Get Pagination
      tags:
      - User
...
  /admin/users/sign-up:
    post:
      consumes:
      - application/json
      description: Create a new User
      parameters:
      - description: Create User request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user_do.CreateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v_api.Response-user_do_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v_proto.VolioRpcError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v_proto.VolioRpcError'
      summary: Create User
      tags:
      - User
swagger: "2.0"