# server/users/users.delete_handler.go
/* !!
 * Code generated by fkit
 * If you need to write additional functions, should create a separate file.
 * File Created: Sunday, 27 April 2025 14:28:41 +07:00
 * Author: VietLD (leduyviet2612@gmail.com)
 * -----
 * Last Modified: Sunday, 27 April 2025 14:28:41 +07:00
 * Modified By: HIEUBV\hieubv
 * -----
 * Copyright 2025. All rights reserved.
 */

package users

import (
	"github.com/bachhieu/fountain/baselib/v_log"
	"github.com/bachhieu/fountain/baselib/v_net/v_api"
	"github.com/bachhieu/fountain/proto/v_proto"
	"github.com/gofiber/fiber/v2"
)

// @Summary      Delete User
// @Description  Delete an existing User
// @Tags         User
// @Accept       json
// @Produce      json
// @Param        Authorization  header    string  true  "Insert your access token" default(access_token)
// @Param        id     query     string  true  "User ID"
// @Success      200          {object}  v_api.Response[string]
// @Failure      400          {object}  v_proto.VolioRpcError
// @Failure      500          {object}  v_proto.VolioRpcError
// @Router       /admin/users [delete]
func (api *UsersAPI) Delete(ctx *fiber.Ctx) error {
	id := ctx.Query("id")
	if id == "" {
		err := v_proto.NewRpcError(int32(v_proto.VolioRpcErrorCodes_BAD_REQUEST), "id is required")
		v_log.V(1).WithError(err).Errorf("UsersAPI::Delete - Error: %+v", err)
		return v_api.WriteError(ctx, err)
	}
	if err := api.usersController.Delete(id); err != nil {
		v_log.V(1).WithError(err).Errorf("UsersAPI::Delete - Error: %+v", err)
		return v_api.WriteError(ctx, err)
	}

	v_log.V(3).Infof("UsersAPI::Delete - Reply: Oke")

	return v_api.WriteSuccess(ctx, "Deleted successfully")
}