# server/users/users.get_handler.go
/* !!
 * Code generated by fkit
 * If you need to write additional functions, should create a separate file.
 * File Created: Sunday, 27 April 2025 14:28:41 +07:00
 * Author: VietLD (leduyviet2612@gmail.com)
 * -----
 * Last Modified: Sunday, 27 April 2025 14:28:41 +07:00
 * Modified By: HIEUBV\hieubv
 * -----
 * Copyright 2025. All rights reserved.
 */

package users

import (
	"github.com/bachhieu/fountain/baselib/base"
	"github.com/bachhieu/fountain/baselib/v_log"
	"github.com/bachhieu/fountain/baselib/v_net/v_api"
	"github.com/bachhieu/fountain/proto/v_proto"
	"github.com/bachhieu/test/biz/dal/do/user_do"
	"github.com/gofiber/fiber/v2"
)

// @Summary      Get User
// @Description  Get list of User with optional filters
// @Tags         User
// @Accept       json
// @Produce      json
// @Param        Authorization  header    string  true   "Insert your access token" default(access_token)
// @Param        id     query     string  true  "ID of the User"
// @Success      200          {object}  v_api.Response[user_do.UserResponse]
// @Failure      400          {object}  v_proto.VolioRpcError
// @Failure      500          {object}  v_proto.VolioRpcError
// @Router       /admin/users/detail [get]
func (api *UsersAPI) Get(ctx *fiber.Ctx) error {
	id := ctx.Query("id", "")
	if id == "" {
		return v_api.WriteError(ctx, v_proto.NewRpcError(int32(v_proto.VolioRpcErrorCodes_BAD_REQUEST), "id is required"))
	}
	res := api.usersController.Get(id)

	v_log.V(3).Infof("UsersAPI::Get - Reply: %s", base.JSONDebugDataString(res))

	return v_api.WriteSuccess(ctx, user_do.NewUserResponse(res))
}