# server/users/users.pagination_handler.go
/* !!
 * Code generated by fkit
 * If you need to write additional functions, should create a separate file.
 * File Created: Sunday, 27 April 2025 14:28:41 +07:00
 * Author: VietLD (leduyviet2612@gmail.com)
 * -----
 * Last Modified: Sunday, 27 April 2025 14:28:41 +07:00
 * Modified By: HIEUBV\hieubv
 * -----
 * Copyright 2025. All rights reserved.
 */

package users

import (
	"github.com/bachhieu/fountain/baselib/base"
	"github.com/bachhieu/fountain/baselib/v_log"
	"github.com/bachhieu/fountain/baselib/v_net/v_api"
	"github.com/bachhieu/fountain/proto/v_proto"
	"github.com/bachhieu/test/biz/dal/do/user_do"
	"github.com/gofiber/fiber/v2"
)

// @Summary      Get Pagination
// @Description  Get pagination of User
// @Tags         User
// @Accept       json
// @Produce      json
// @Param        Authorization  header    string  true  "Insert your access token" default(access_token)
// @Param        query_params              query     user_do.UserQueryParams  false  "Query params"
// @Success      200          {object}  v_api.Response[[]user_do.UserResponse]
// @Failure      400          {object}  v_proto.VolioRpcError
// @Failure      500          {object}  v_proto.VolioRpcError
// @Router       /admin/users/pagination [get]
func (api *UsersAPI) GetPagination(ctx *fiber.Ctx) error {
	params := new(user_do.UserQueryParams)
	if err := ctx.QueryParser(params); err != nil {
		v_log.V(1).WithError(err).Errorf("UsersAPI::GetPagination - Parse params error: %+v", err)
		return v_api.WriteError(ctx, v_proto.NewRpcError(int32(v_proto.VolioRpcErrorCodes_BAD_REQUEST), "invalid parameters"))
	}

	offset, limit := v_api.GetQueryOffsetLimit(ctx)
	res, err := api.usersController.Gets(params, offset, limit)
	if err != nil {
		v_log.V(1).WithError(err).Errorf("UsersAPI::GetPagination - Error: %+v", err)
		return v_api.WriteError(ctx, err)
	}

	total := api.usersController.CountTotal(params)

	v_log.V(3).Infof("UsersAPI::GetPagination - Reply: %s", base.JSONDebugDataString(res))

	return v_api.WithContext(ctx).WithPaginationByLoadMore(total, offset, limit).WriteSuccess(user_do.NewUserResponses(res))
}